Reorganizing and extending words embedding docs - to DEV (#13)
* updating words embeddings docs

* removeing we.core

* uniting demos (notebooks) for words embedding to one

* changing box about superficial effect to note

* phrasing description of link to demo

* polishing documentation
