+++ docs/demos.rst
@@ -7,6 +7,4 @@ Demos
 
   notebooks/demo-compas-analysis
   notebooks/demo-fico-analysis
-  notebooks/demo-gender-bias-words-embedding
-  notebooks/exploratory-demo-race-bias-words-embedding
-  notebooks/demo-weat-alpha
+  notebooks/demo-words-embedding-bias
