+++ ethically/we/benchmark.py
@@ -1,10 +1,12 @@
 """
-Evaluate words embeedings by standard benchmarks.
+Evaluate words embedding by standard benchmarks.
 
-Reference: https://github.com/kudkudak/word-embeddings-benchmarks
+Reference:
+- https://github.com/kudkudak/word-embeddings-benchmarks
 
 
 Word Pairs Tasks
+~~~~~~~~~~~~~~~~
 
 1. The WordSimilarity-353 Test Collection
    http://www.cs.technion.ac.il/~gabr/resources/data/wordsim353/
@@ -29,6 +31,7 @@ Word Pairs Tasks
 
 
 Analogies Tasks
+~~~~~~~~~~~~~~~
 
 1. Google Analogies (subset of WordRep)
    https://code.google.com/archive/p/word2vec/source
@@ -60,13 +63,13 @@ PAIR_WORDS_EVALUATION_FIELDS = ['pearson_r', 'pearson_pvalue',
                                 'ratio_unkonwn_words']
 
 
-def get_data_resource_path(filename):
+def _get_data_resource_path(filename):
     return resource_filename(__name__, os.path.join('data',
                                                     'benchmark',
                                                     filename))
 
 
-def prepare_word_pairs_file(src, dst, delimiter='\t'):
+def _prepare_word_pairs_file(src, dst, delimiter='\t'):
     """Transform formats of word pairs files to tsv."""
     df = pd.read_csv(src, header=None, delimiter=delimiter)
     df.loc[:, :2].to_csv(dst, sep=delimiter, index=False, header=False)
@@ -81,7 +84,7 @@ def evaluate_word_pairs(model, kwargs_word_pairs=None):
                               evaluate_word_pairs
                               method.
     :type kwargs_word_pairs: dict or None
-    :return: DataFrame of evaluation results.
+    :return: :class:`pandas.DataFrame` of evaluation results.
     """
 
     if kwargs_word_pairs is None:
@@ -90,7 +93,7 @@ def evaluate_word_pairs(model, kwargs_word_pairs=None):
     results = {}
 
     for name, filename in WORD_PAIRS_TASKS.items():
-        path = get_data_resource_path(filename)
+        path = _get_data_resource_path(filename)
         (pearson,
          spearman,
          ratio_unknown_words) = model.evaluate_word_pairs(path,
@@ -119,7 +122,7 @@ def evaluate_word_analogies(model, kwargs_word_analogies=None):
                                   evaluate_word_analogies
                                   method.
     :type evaluate_word_analogies: dict or None
-    :return: DataFrame of evaluation results.
+    :return: :class:`pandas.DataFrame` of evaluation results.
     """
 
     if kwargs_word_analogies is None:
@@ -128,7 +131,7 @@ def evaluate_word_analogies(model, kwargs_word_analogies=None):
     results = {}
 
     for name, filename in ANALOGIES_TASKS.items():
-        path = get_data_resource_path(filename)
+        path = _get_data_resource_path(filename)
         overall_score, _ = model.evaluate_word_analogies(path,
                                                          **kwargs_word_analogies)  # pylint: disable=C0301
